refactor: rename hook useUserProjects to useProjects
